HW02 - UserMeal Spring realization

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryUserMealRepositoryImpl.java
@@ -1,18 +1,22 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.springframework.core.CollectionFactory;
+import org.springframework.stereotype.Repository;
+import ru.javawebinar.topjava.LoggedUser;
 import ru.javawebinar.topjava.model.UserMeal;
 import ru.javawebinar.topjava.repository.UserMealRepository;
 import ru.javawebinar.topjava.util.UserMealsUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 /**
  * GKislin
  * 15.09.2015.
  */
+@Repository
 public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
     private Map<Integer, UserMeal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
@@ -31,18 +35,30 @@ public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
     }
 
     @Override
-    public void delete(int id) {
+    public boolean delete(int id) {
+        UserMeal meal = repository.get(id);
+        if (meal == null || meal.getUserId() != LoggedUser.id())
+            return false;
         repository.remove(id);
+        return true;
     }
 
     @Override
     public UserMeal get(int id) {
+        UserMeal meal = repository.get(id);
+        if (meal == null || meal.getUserId() != LoggedUser.id())
+            return null;
         return repository.get(id);
     }
 
     @Override
     public Collection<UserMeal> getAll() {
-        return repository.values();
+        Collection<UserMeal> meals = repository.values();
+        List<UserMeal> mealList = meals.stream()
+                .filter(userMeal -> userMeal.getUserId() == LoggedUser.id())
+                .collect(Collectors.toList());
+        Collections.sort(mealList, Comparator.comparing(UserMeal::getDateTime).reversed());
+        return mealList;
     }
 }
